flask_app/requirements.txt
Flask==3.0.3
flask-cors
mysql-connector-python
firebase-admin==5.0.3